--- src/button.css
@@ -1,8 +1,0 @@
-.button-box {
-  width: 100%;
-  text-align: right;
-}
-
-.button-text {
-  margin-right: 10px;
-}
