+++ src/index.css
@@ -1,4 +1,3 @@
-@import './button.css';
 @import './utils.css';
 
 body {
